WIP: src/ksm/discovery.go: refactor into generic endpoints discoverer
And specific KSM discoverer. This separates KSM-specific logic away from
generic Kubernetes logic.
